corniscan/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.31.0",
    "ezdxf>=1.4.2",
    "fastapi>=0.128.8",
    "opencv-python-headless>=4.13.0.92",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "resend>=2.22.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.39.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
